+++ src/lib.rs
@@ -1,9 +1,12 @@
 mod ssh;
+mod service;
 
 use ed25519_dalek::{PUBLIC_KEY_LENGTH, SECRET_KEY_LENGTH};
 use iroh::{protocol::Router, Endpoint};
 
 pub use ssh::dot_ssh;
+pub use service::install_service;
+pub use service::ServiceParams;
 
 
 #[derive(Debug, Clone)]
