+++ src/main.rs
@@ -3,8 +3,8 @@ use std::str::FromStr;
 use anyhow::bail;
 use clap::{Parser, Subcommand, command};
 use iroh::{NodeId, SecretKey};
-use iroh_ssh::{IrohSsh, dot_ssh};
-use tokio::process::Command;
+use iroh_ssh::{dot_ssh, install_service, IrohSsh, ServiceParams};
+
 
 #[derive(Parser)]
 #[command(name = "irohssh", about = "SSH without IP", after_help = "
@@ -84,53 +84,9 @@ async fn info_mode() -> anyhow::Result<()> {
 }
 
 async fn service_mode(ssh_port: u16) -> anyhow::Result<()> {
-    // only on linux with systemctl
-    if !cfg!(target_os = "linux") {
-        anyhow::bail!("Service mode is only supported on linux")
+    if install_service(ServiceParams { ssh_port }).is_err() {
+        println!("Service mode is only supported on linux");
     }
-
-    let service_raw = r#"[Unit]
-Description=SSH over Iroh
-
-[Service]
-Type=simple
-WorkingDirectory=~
-ExecStart=/bin/bash -c 'iroh-ssh server -p --ssh-port [SSHPORT]'
-Restart=on-failure
-RestartSec=3s
-
-[Install]
-WantedBy=multi-user.target
-"#;
-
-    let service_raw = service_raw.replace("[SSHPORT]", &ssh_port.to_string());
-
-    let service_path = std::path::Path::new("/etc/systemd/system/iroh-ssh-server.service");
-    std::fs::write(service_path, service_raw)?;
-
-    // check if service is started and running and print status
-    let status = Command::new("systemctl")
-        .arg("is-active")
-        .arg("iroh-ssh-server.service")
-        .output()
-        .await?;
-
-    if status.status.success() {
-        println!("Service is already running");
-    } else {
-        println!("Starting service...");
-        Command::new("systemctl")
-            .arg("enable")
-            .arg("iroh-ssh-server.service")
-            .output()
-            .await?;
-        Command::new("systemctl")
-            .arg("start")
-            .arg("iroh-ssh-server.service")
-            .output()
-            .await?;
-    }
-
     Ok(())
 }
